# waspadai-product-docs/contracts/future/product-api.openapi.yaml
...
components:
  schemas:
    AIResult:
      $ref: '#/components/schemas/VerificationResponse'
...
    CommunityDetail:
      additionalProperties: false
      properties:
        case_id:
          format: uuid
          type: string
        counts:
          additionalProperties: false
          properties:
            DIBANTAH:
              minimum: 0
              type: integer
            DIDUKUNG:
              minimum: 0
              type: integer
          required:
          - DIDUKUNG
          - DIBANTAH
          type: object
        published_at:
          format: date-time
          type: string
        redacted_image_url:
          anyOf:
          - format: uri
            type: string
          - type: 'null'
        redacted_text:
          type: string
        result:
          $ref: '#/components/schemas/AIResult'
        status:
          enum:
          - PUBLISHED_UNVERIFIED
          - VERIFIED_EVIDENCE
          type: string
        title:
          type: string
        user_vote:
          anyOf:
          - enum:
            - DIDUKUNG
            - DIBANTAH
            type: string
          - type: 'null'
        execution_mode:
          type: string
          enum:
          - REMOTE
          - MOCK
          description: Product execution provenance; synthetic fixture has MOCK even
            though strict upstream result.mode is LIVE.
      required:
      - case_id
      - title
      - redacted_text
      - status
      - published_at
      - counts
      - redacted_image_url
      - result
      - user_vote
      - execution_mode
      type: object
    CommunityItem:
      additionalProperties: false
      properties:
        case_id:
          format: uuid
          type: string
        counts:
          additionalProperties: false
          properties:
            DIBANTAH:
              minimum: 0
              type: integer
            DIDUKUNG:
              minimum: 0
              type: integer
          required:
          - DIDUKUNG
          - DIBANTAH
          type: object
        published_at:
          format: date-time
          type: string
        redacted_text:
          type: string
        status:
          enum:
          - PUBLISHED_UNVERIFIED
          - VERIFIED_EVIDENCE
          type: string
        title:
          type: string
      required:
      - case_id
      - title
      - redacted_text
      - status
      - published_at
      - counts
      type: object
    CommunityPage:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/CommunityItem'
          type: array
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - items
      - next_cursor
      type: object
    CommunityState:
      additionalProperties: false
      properties:
        case_id:
          format: uuid
          type: string
        community_state:
          enum:
          - PRIVATE
          - PUBLISHED_UNVERIFIED
          - VERIFIED_EVIDENCE
          - WITHDRAWN
          type: string
        revision:
          minimum: 1
          type: integer
      required:
      - case_id
      - community_state
      - revision
      type: object
...
    Error:
      additionalProperties: false
      properties:
        error:
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              format: uuid
              type: string
            retry_after_seconds:
              anyOf:
              - minimum: 0
                type: integer
              - type: 'null'
            retryable:
              type: boolean
          required:
          - code
          - message
          - request_id
          - retryable
          - retry_after_seconds
          type: object
      required:
      - error
      type: object
...
    Preview:
      additionalProperties: false
      properties:
        confirmation_required:
          enum:
          - true
          type: boolean
        expires_at:
          format: date-time
          type: string
        preview_id:
          format: uuid
          type: string
        caption:
          minLength: 1
          maxLength: 5000
          type: string
        redacted_image_url:
          anyOf:
          - format: uri
            type: string
          - type: 'null'
        redacted_text:
          type: string
        redactions:
          items:
            type: string
          type: array
      required:
      - preview_id
      - expires_at
      - redacted_text
      - redacted_image_url
      - redactions
      - confirmation_required
      type: object
...
    PublishRequest:
      additionalProperties: false
      properties:
        publication_consent:
          enum:
          - true
          type: boolean
        rag_reuse_consent:
          type: boolean
        preview_id:
          format: uuid
          type: string
      required:
      - preview_id
      - publication_consent
      - rag_reuse_consent
      - caption
      type: object
...
    VoteRequest:
      additionalProperties: false
      properties:
        vote:
          enum:
          - DIDUKUNG
          - DIBANTAH
          type: string
      required:
      - vote
      type: object
    VoteResult:
      additionalProperties: false
      properties:
        case_id:
          format: uuid
          type: string
        counts:
          additionalProperties: false
          properties:
            DIBANTAH:
              minimum: 0
              type: integer
            DIDUKUNG:
              minimum: 0
              type: integer
          required:
          - DIDUKUNG
          - DIBANTAH
          type: object
        user_vote:
          anyOf:
          - enum:
            - DIDUKUNG
            - DIBANTAH
            type: string
          - type: 'null'
      required:
      - case_id
      - user_vote
      - counts
      type: object
...
  securitySchemes:
    SupabaseBearer:
      bearerFormat: JWT
      description: Validated Supabase access token; no internal AI key on clients.
      scheme: bearer
      type: http
  responses:
    SafeError:
      description: Safe Product error envelope; code/action defined in integration
        guide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
info:
  description: Product target contract aligned to user-uploaded WaspadAI AI OpenAPI
    0.7.0. All upstream schemas copied exactly. Product workflows remain draft; GitHub
    commit and live deployment not verified.
  title: WaspadAI Product API — DRAFT
  version: 3.0.1-draft
openapi: 3.1.0
paths:
...
  /api/v1/community:
    get:
      operationId: get_api_v1_community
      parameters:
      - in: query
        name: limit
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - in: query
        name: cursor
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPage'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Community GET
      tags:
      - Community
      x-implementation-status: TARGET_DRAFT
  /api/v1/community/{case_id}:
    get:
      operationId: get_api_v1_community_case_id
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityDetail'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Community GET
      tags:
      - Community
      x-implementation-status: TARGET_DRAFT
  /api/v1/community/{case_id}/vote:
    delete:
      operationId: delete_api_v1_community_case_id_vote
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResult'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Community DELETE
      tags:
      - Community
      x-implementation-status: TARGET_DRAFT
    post:
      operationId: post_api_v1_community_case_id_vote
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResult'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: PROPOSED POST atomic vote upsert; older source used PUT, confirm before
        freeze
      tags:
      - Community
      x-implementation-status: TARGET_DRAFT
...
  /api/v1/history/{case_id}/community:
    delete:
      operationId: delete_api_v1_history_case_id_community
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: Completed; no response body
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Community DELETE
      tags:
      - Community
      x-implementation-status: TARGET_DRAFT
    post:
      operationId: post_api_v1_history_case_id_community
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Product-local idempotency, not a guarantee of upstream exactly-once
          inference.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityState'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Community POST
      tags:
      - Community
      x-implementation-status: TARGET_DRAFT
  /api/v1/history/{case_id}/community-preview:
    post:
      operationId: post_api_v1_history_case_id_community_preview
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Product-local idempotency, not a guarantee of upstream exactly-once
          inference.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preview'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Community POST
      tags:
      - Community
      x-implementation-status: TARGET_DRAFT
...
servers:
- description: Local Product target only
  url: http://127.0.0.1:8001
...